feat(api): Added optional `longError` option to `isValid()`
Depending on use, violations reportd by isValid are either serialized or used as an object. This
flag provides caller control on how much information is put into the error message, allowing for
minimal and clean errors when the errors are being processed, and convenience in dumping all
violations to console when that is more appropriate. Turns out, this code was being repeated in each
use of the API. This refactoring just mirrors real use.